don't import modification_timestamp from chapter fragments files
That's because Notabenoid doesn't store modification timestamps as I
initially thought. The only timestamp that is stored refers to when a
fragment was created and thus is not updated when someone edits that
fragment. Of course, I could match changed fragments with the overall
chapter modification timestamps as reported on the main page of the
translation, but first of all, this is a concern of the scraper program,
and secondly - who cares.