get rid of memory mapped file I/O
I do this because mmapping didn't live up to its performance promises to
the point that the difference was next to none (primarily because reads
and writes are relatively sequential in my use-case). Instead, mmap only
complicated code paths and had very dubious safety guarantees, in that
it was unclear whether Rustc's assumptions for optimizing based on
mutability of references would lead to runtime UB if the file was
changed while being read, if it all. Usually discussions of mmap UB in
Rust conclude with handwaving that "it is safe unless the file is
modified out-of-process", which doesn't exactly help.

This reverts commit 7176cd3d2fb7b4c200ac02d475f4e360a98cd754