implement Deserialize and Serialize properly for backend::Message
They used to be implemented in local methods because I wanted to
integrate method dispatch into serialization and deserialization
process. Unfortunately, due to the fact that these traits are generic,
it is hard to integrate them into the dispatcher since it has to work
with regular function pointers, and function pointers can't have generic
parameters. Maybe serde_erased allows doing this, but I don't know. So,
I leave this approach for another day.